Moved UserIndicator to service-statistics
