+++ control-statistics/src/test/resources/fi/nls/oskari/control/statistics/user/GetUserIndicatorsHandlerTest-id5.json
@@ -17,11 +17,5 @@
   },
   "public": true,
   "year": 2013,
-  "data": [
-    {
-      "region": 49,
-      "primary value": 42
-    }
-  ],
   "category": "KUNTA"
 }
